readded not silly refernce removed really silly refernce some tests with poll()

+++ tunDevice.h
@@ -43,7 +43,7 @@ public:
   void close();
   bool isOpen();
 
-  int read(Buffer& buf);
+  short read(Buffer& buf);
   int write(Buffer& buf);
 
   char* getActualName();
